chore: deprecate cli use cases
